+++ changelog.md
@@ -1,3 +1,20 @@
+# <sub>v2.0.0-alpha</sub>
+#### _Nov. 26, 2017_
+
+  * Rename `getEmptyObjectData` to `getEmptyDataObject`.
+  * Rename `parseObjectData` to `parseDataObject`.
+  * Rename `processStringParam` to `parseString`.
+  * Rename `processIntParam` to `parseInt`.
+  * Rename `processFloatParam` to `parseFloat`.
+  * Rename `processDateParam` to `parseDate`.
+  * Rename `processId` to `parseId`.
+  * Rename `processIdList` to `parseIdList`.
+  * Rename `processObjectId` to `parseObjectId`.
+  * Rename `processIn` to `parseIn`.
+  * Rename `processNin` to `parseNin`.
+  * Rename `processFields` to `parseFields`.
+  * Remove `processFilterByName`.
+
 # <sub>v1.1.0</sub>
 #### _Nov. 25, 2017_